refactored identicon generator
