+++ app/.meteor/packages
@@ -32,3 +32,7 @@ splendido:accounts-meld
 london:body-class
 reactive-dict
 agnito:simptip
+meteorhacks:npm
+
+
+npm-container
