+++ app/.meteor/versions
@@ -53,10 +53,12 @@ logging@1.0.7
 london:body-class@2.1.0
 meteor@1.1.6
 meteor-platform@1.2.2
+meteorhacks:async@1.0.0
 meteorhacks:fast-render@2.3.2
 meteorhacks:flow-layout@1.1.1
 meteorhacks:flow-router@1.4.1
 meteorhacks:inject-data@1.2.3
+meteorhacks:npm@1.3.0
 meteorhacks:picker@1.0.2
 meteorhacks:subs-manager@1.3.0
 meteortoys:toykit@0.3.4
@@ -67,6 +69,7 @@ momentjs:moment@2.10.3
 mongo@1.1.0
 msavin:mongol@1.0.22
 npm-bcrypt@0.7.8_2
+npm-container@1.0.0
 oauth@1.1.4
 oauth2@1.1.3
 observe-sequence@1.0.6
