+++ app/packages.json
@@ -1,0 +1,3 @@
+{
+  "retricon": "0.0.3"
+}
